- bugFix: Search namespace renamed to User by mistake reverted
